Move renderer to own file. And use Raw for amp (for now)
