--- frontend/components/elements/ElementRenderer.tsx
@@ -1,52 +1,0 @@
-import { RawComponent } from '../elements/RawComponent';
-import { ImageBlockComponent } from './ImageBlockComponent/amp';
-import React from 'react';
-import { css } from 'emotion';
-export const ArticleRenderer: React.SFC<{ elements: CAPIElement[] }> = ({
-    elements,
-}) => {
-    const output = elements
-        .map((element, i) => {
-            switch (element._type) {
-                case 'model.liveblog.TextBlockElement':
-                    return <RawComponent key={i} html={element.html} />;
-                case 'model.liveblog.TweetBlockElement':
-                    return (
-                        <div
-                            className={css`
-                                width: 100%;
-                                background: aqua;
-                            `}
-                        >
-                            <RawComponent key={i} html={element.html} />
-                        </div>
-                    );
-                case 'model.liveblog.ImageBlockElement':
-                    return <ImageBlockComponent key={i} element={element} />;
-                default:
-                    return null;
-            }
-        })
-        .filter(_ => _ != null);
-    return <>{output}</>;
-};
-
-export const AMPRenderer: React.SFC<{ elements: CAPIElement[] }> = ({
-    elements,
-}) => {
-    const output = elements
-        .map((element, i) => {
-            switch (element._type) {
-                case 'model.liveblog.TextBlockElement':
-                    return <RawComponent key={i} html={element.html} />;
-                case 'model.liveblog.TweetBlockElement':
-                    return <RawComponent key={i} html={element.html} />;
-                case 'model.liveblog.ImageBlockElement':
-                    return <ImageBlockComponent key={i} element={element} />;
-                default:
-                    return null;
-            }
-        })
-        .filter(_ => _ != null);
-    return <>{output}</>;
-};
