+++ frontend/components/elements/ImageBlockComponent/amp.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { RawComponent } from '../RawComponent';
 
 const getFallback: (images: Image[]) => Image = images =>
     images.find(_ => _.fields.isMaster === 'true') || images[0];
@@ -8,11 +9,12 @@ export const ImageBlockComponent: React.SFC<{ element: ImageBlockElement }> = ({
 }) => {
     const image = getFallback(element.media.allImages);
     return (
-        <amp-img
-            src={image.url}
-            alt={element.data.alt}
-            height={image.fields.height}
-            width={image.fields.width}
+        <RawComponent
+            html={`<amp-img src="${image.url}" alt="${
+                element.data.alt
+            }" height="${image.fields.height}" width="${
+                image.fields.width
+            }" />`}
         />
     );
 };
